+++ src/tools/ufo2map/csg.c
@@ -104,8 +104,8 @@ static qboolean BrushesDisjoint (bspbrush_t *a, bspbrush_t *b)
 	return qfalse;	/* might intersect */
 }
 
-static int minplanenums[2];
-static int maxplanenums[2];
+static uint16_t minplanenums[2];
+static uint16_t maxplanenums[2];
 
 /**
  * @brief Any planes shared with the box edge will be set to no texinfo
